Crate media_types

Media Types also known as MIME types describe the nature of data they are used in email to describe the file type of attachments and in HTTP to to give the type of a resource.

There are many RFCs describing media types the two most important for this crate is RFC 2046 - Multipurpose Internet Mail Extensions (MIME) Part Two: Media Types.

Reexports

pub use self::Type::{Application, Audio, Image, Message, Model, Multipart, Text, Video};

pub use self::Tree::{Personal, Private, Standards, Vendor};

Structs

MediaType
A Media Type commonly used to describe the contents of a resource.

Enums

Charset
A Mime charset.

Error
Defines an Error type for media types.

Tree
Provides the four registration trees.

Type
Provides the six discrete and the two composite top-level media types.

Type Definitions

Result
Result type used for this library.