Enum media_types::Tree

pub enum Tree {
    Standards,
    Vendor,
    Personal,
    Private,
    Unregistered(Cow<'static, str>),
}

Provides the four registration trees.

Variants

Standards

The standards tree is intended for types of general interest to the Internet community.

Vendor

The vendor tree is used for media types associated with publicly available products.

Personal

Registrations for media types created experimentally or as part of products that are not distributed commercially may be registered in the personal or vanity tree.

Private

Subtype names with "x." as the first facet may be used for types intended exclusively for use in private, local environments.

Unregistered

Other unofficial trees.