Enum media_types::Type

pub enum Type {
    Text,
    Image,
    Audio,
    Video,
    Application,
    Multipart,
    Message,
    Model,
    Unregistered(Cow<'static, str>),
}

Provides the six discrete and the two composite top-level media types.

Variants

Text

The "text" top-level type is intended for sending material that is principally textual in form.

Image

A top-level type of "image" indicates that the content specifies one or more individual images.

Audio

A top-level type of "audio" indicates that the content contains audio data.

Video

A top-level type of "video" indicates that the content specifies a time-varying-picture image, possibly with color and coordinated sound.

Application

The "application" top-level type is to be used for discrete data that do not fit under any of the other type names, and particularly for data to be processed by some type of application program.

Multipart

The "multipart" top-level type is to be used for data consisting of multiple entities of independent data types.

Message

A body of media type "message" is itself all or a portion of some kind of message object.

Model

The "model" media type is used for 3D-models.

Unregistered

Less common top-level types.