Module media_types::type_

Provides the five discrete and the two composite top-level media types.

Constants

APPLICATION
The "application" top-level type is to be used for discrete data that do not fit under any of the other type names, and particularly for data to be processed by some type of application program.

AUDIO
A top-level type of "audio" indicates that the content contains audio data.

IMAGE
A top-level type of "image" indicates that the content specifies one or more individual images.

MESSAGE
A body of media type "message" is itself all or a portion of some kind of message object.

MULTIPART
The "multipart" top-level type is to be used for data consisting of multiple entities of independent data types..

TEXT
The "text" top-level type is intended for sending material that is principally textual in form.

VIDEO
A top-level type of "video" indicates that the content specifies a time-varying-picture image, possibly with color and coordinated sound.